+++ engine/SceneServer/PlayerEntity.h
@@ -7,11 +7,11 @@
 class PlayerEntity : public SceneEntity {
 private:
 	int m_connId;
-	int m_roleId;
+	int64_t m_roleId;
 	std::string m_name;
 	void* m_scriptObj;
 public:
-	PlayerEntity(int connId, int actorId, int roleId, std::string name, int x, int y, int moveSpeed, void* gameScene, GridChgFunc posChgFunc);
+	PlayerEntity(int connId, int actorId, int64_t roleId, std::string name, int x, int y, int moveSpeed, void* gameScene, GridChgFunc posChgFunc);
 	inline void setScriptObj(void* scriptObj) { m_scriptObj = scriptObj; }
 
 	inline int getConnId() { return m_connId; }
